# fix: avoid resetting `method.hashUpdate`