--- a/sources/ippcp/hash/sha224/pcphashmethod_sha224_tt.c
+++ b/sources/ippcp/hash/sha224/pcphashmethod_sha224_tt.c
@@ -46,29 +46,32 @@
 //
 *F*/
 
-IPPFUN( const IppsHashMethod*, ippsHashMethod_SHA224_TT, (void) )
+IPPFUN(const IppsHashMethod*, ippsHashMethod_SHA224_TT, (void))
 {
-   static IppsHashMethod method = {
-      ippHashAlg_SHA224,
-      IPP_SHA224_DIGEST_BITSIZE/8,
-      MBS_SHA256,
-      MLR_SHA256,
-      NULL,
-      NULL,
-      NULL,
-      NULL,
-   };
+    static IppsHashMethod method = {
+        ippHashAlg_SHA224,
+        IPP_SHA224_DIGEST_BITSIZE / 8,
+        MBS_SHA256,
+        MLR_SHA256,
+        NULL,
+        NULL,
+        NULL,
+        NULL,
+    };
 
-   // don't merge `method` initialization with function pointers assignment
-   // to prevent relocations (indirect calls) to be generated in the binary
-   method.hashInit   = sha224_hashInit;
-   method.hashUpdate = sha256_hashUpdate;
-   method.hashOctStr = sha224_hashOctString;
-   method.msgLenRep  = sha256_msgRep;
+    // don't merge `method` initialization with function pointers assignment
+    // to prevent relocations (indirect calls) to be generated in the binary
+    method.hashInit   = sha224_hashInit;
+    method.hashOctStr = sha224_hashOctString;
+    method.msgLenRep  = sha256_msgRep;
 
-   #if (_SHA_NI_ENABLING_==_FEATURE_TICKTOCK_ || _SHA_NI_ENABLING_==_FEATURE_ON_)
-   if(IsFeatureEnabled(ippCPUID_SHA))
-      method.hashUpdate = sha256_ni_hashUpdate;
-   #endif
-   return &method;
+#if (_SHA_NI_ENABLING_ == _FEATURE_TICKTOCK_ || _SHA_NI_ENABLING_ == _FEATURE_ON_)
+    if (IsFeatureEnabled(ippCPUID_SHA)) {
+        method.hashUpdate = sha256_ni_hashUpdate;
+        return &method;
+    }
+#endif
+
+    method.hashUpdate = sha256_hashUpdate;
+    return &method;
 }