This is done to avoid setting the variable twice, because in a multithreaded environment this destroys memory caches. Note that to make the code cleaner I used an early return in the special `sha256_ni_hashUpdate` case.

Compilation using the Intel® oneAPI DPC++/C++ compiler succeeded.
Code formatted according to the [.clang-format](.clang-format) rules.
Added a note in change log and incremented the version.

Fixes intel/linux-sgx#1073