@mikolajlubiak 2025-03-04
Did a benchmark and the new code saves about 1,000,000,000 cache references and 10 CPU seconds in the specific benchmark case.
New:
```
❯ sudo perf stat -B -e cache-references,cache-misses,cycles,instructions,branches,faults,migrations ./benchmark 12 10000000
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 5652985402 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 5654593436 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6133624034 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6137065569 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6217676274 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6218518712 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6277306025 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6278381844 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6356617497 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6360215651 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6448754165 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6455433524 ns

 Performance counter stats for './benchmark 12 10000000':

     2,539,162,050      cache-references
            97,644      cache-misses                     #    0.00% of all cache refs
   303,215,955,461      cycles
    56,939,106,874      instructions                     #    0.19  insn per cycle
     5,594,461,691      branches
               162      faults
                 9      migrations

       6.457206156 seconds time elapsed

      73.972820000 seconds user
       0.005963000 seconds sys
```
Old:
```
❯ sudo perf stat -B -e cache-references,cache-misses,cycles,instructions,branches,faults,migrations ./benchmark 12 10000000
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6324792106 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6328368805 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6948123032 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6972823042 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6989185139 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 7003213614 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 7044943411 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 7054015430 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 7101390258 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 7104537262 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 7106029344 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 7121662468 ns

 Performance counter stats for './benchmark 12 10000000':

     3,727,422,918      cache-references
           120,280      cache-misses                     #    0.00% of all cache refs
   339,299,104,432      cycles
   110,108,473,947      instructions                     #    0.32  insn per cycle
     5,839,178,741      branches
               164      faults
                10      migrations

       7.124216913 seconds time elapsed

      82.828447000 seconds user
       0.005983000 seconds sys
```