@mikolajlubiak 2025-03-04
Some other stats with the same benchmark:

With fix:
```
❯ sudo perf stat -B -e LLC-loads,LLC-load-misses,LLC-stores,LLC-store-misses ./benchmark 12 10000000
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6149230421 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6153150286 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6195275445 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6204341769 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6204890586 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6208547958 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6266405663 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6269321465 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6269740576 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6274113901 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6402036998 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6408649862 ns

 Performance counter stats for './benchmark 12 10000000':

       289,679,436      LLC-loads                                                               (50.00%)
            17,378      LLC-load-misses                  #    0.01% of all LL-cache accesses    (49.99%)
       256,823,565      LLC-stores                                                              (50.01%)
             2,493      LLC-store-misses                                                        (50.01%)

       6.410538773 seconds time elapsed

      74.650541000 seconds user
       0.000993000 seconds sys
```

Without fix:
```
❯ sudo perf stat -B -e LLC-loads,LLC-load-misses,LLC-stores,LLC-store-misses ./benchmark 12 10000000
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6249371647 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6254551471 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6717360964 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6739020889 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6950421995 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 6963391824 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 7013288374 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 7060327093 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 7068903938 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 7072578410 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 7175201365 ns
SHA256 result: 6b4ab10d92373474a97d10639e667f6b734af012f9be29997f1befd1c7166199 in 7182094432 ns

 Performance counter stats for './benchmark 12 10000000':

       467,841,050      LLC-loads                                                               (50.00%)
            18,428      LLC-load-misses                  #    0.00% of all LL-cache accesses    (49.99%)
       275,681,401      LLC-stores                                                              (50.01%)
             2,539      LLC-store-misses                                                        (50.01%)

       7.183901829 seconds time elapsed

      82.007108000 seconds user
       0.014439000 seconds sys
```